+++ plugin/resources/assets/js/pages/instanceForm/sections/DeadlineSection.vue
@@ -16,11 +16,18 @@
 
             <button type="button" class="add-deadline-btn" @click="onAddDeadlineClicked">Add Deadline</button>
 
-            <br>
+            <br />
 
             <label v-if="isEditing">
                 <input type="checkbox" name="recalculate_grades" v-model="form.recalculate_grades">
                 Recalculate grades
+            </label>
+            <br />
+            <label v-if="isEditing && form.recalculate_grades === true && form.fields.grading_method_code === 3"
+                   class="red--text">
+
+                Warning: recalculating grades with grading method 'prefer_best_each_test_grade'
+                might take a lot of time to finish.
             </label>
 
         </div>